pyproject.toml
[tool.poetry]
name = "bookstore"
version = "0.1.0"
description = "Bookstore API"
authors = ["Helbert guirra <helbert.guirra@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
pytest = "^9.0.2"
factory-boy = "^3.3.3"
django = "^6.0.1"
djangorestframework = "^3.16.1"
django-rest-framework = "^0.1.0"
django-extensions = "^4.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"